+++ tests/conftest.py
@@ -1,5 +1,7 @@
 import pytest
 
+pytest_plugins = ["ome_zarr_converters_tools.testing.plugin"]
+
 
 @pytest.fixture(autouse=True)
 def set_env_vars(monkeypatch: pytest.MonkeyPatch) -> None: